+++ plone/app/caching/tests/test_purge.py
@@ -37,6 +37,7 @@ from zope.lifecycleevent import ObjectModifiedEvent
 from zope.lifecycleevent import ObjectMovedEvent
 from zope.lifecycleevent import ObjectRemovedEvent
 
+import six
 import unittest
 
 if HAVE_AT:
@@ -448,9 +449,14 @@ class TestScalesPurgePaths(unittest.TestCase):
 
     def test_scale_purge_paths_unicode(self):
         purge = ScalesPurgePaths(self.file)
-        self.assertEqual(
+        expected = [
+            u'/plone/media/file/view/++widget++form.widgets.file/@@download/data/töstfile.csv',  # noqa: E501
+            u'/plone/media/file/@@download/file/data/töstfile.csv',
+        ]
+        if six.PY2:
+            # the getRelativePaths method returns bytes on Python 2
+            expected = [x.encode('utf8') for x in expected]
+        self.assertListEqual(
             list(purge.getRelativePaths()),
-            ['/plone/media/file/view/++widget++form.widgets.file/@@download/' +
-             'data/t\xc3\xb6stfile.csv',
-             '/plone/media/file/@@download/file/data/t\xc3\xb6stfile.csv'],
+            expected,
         )