+++ plone/app/caching/tests/test_profile_without_caching_proxy.py
@@ -135,7 +135,7 @@ class TestProfileWithoutCaching(unittest.TestCase):
         # This should use cacheInBrowser
         self.assertEqual('max-age=0, must-revalidate, private',
                          browser.headers['Cache-Control'])
-        tag = '"|test_user_1_|{0}|en|{1}|0|0'.format(
+        tag = '"|test_user_1_|{0}|en|{1}|0'.format(
             catalog.getCounter(),
             default_skin,
         )
@@ -153,7 +153,7 @@ class TestProfileWithoutCaching(unittest.TestCase):
                          browser.headers['X-Cache-Operation'])
         self.assertEqual('max-age=0, must-revalidate, private',
                          browser.headers['Cache-Control'])
-        tag = '"|test_user_1_|{0}|en|{1}|0|1'.format(
+        tag = '"|test_user_1_|{0}|en|{1}|0'.format(
             catalog.getCounter(),
             default_skin,
         )
@@ -175,7 +175,7 @@ class TestProfileWithoutCaching(unittest.TestCase):
         # This should use cacheInBrowser
         self.assertEqual('max-age=0, must-revalidate, private',
                          browser.headers['Cache-Control'])
-        tag = '"|test_user_1_|{0}|en|{1}|0'.format(
+        tag = '"|test_user_1_|{0}|en|{1}'.format(
             catalog.getCounter(),
             default_skin,
         )
@@ -223,7 +223,7 @@ class TestProfileWithoutCaching(unittest.TestCase):
         # This should use cacheInBrowser
         self.assertEqual('max-age=0, must-revalidate, private',
                          browser.headers['Cache-Control'])
-        tag = '"||{0}|en|{1}|0|0'.format(catalog.getCounter(), default_skin)
+        tag = '"||{0}|en|{1}|0'.format(catalog.getCounter(), default_skin)
         self.assertEqual(tag, _normalize_etag(browser.headers['ETag']))
         self.assertGreater(now, dateutil.parser.parse(
             browser.headers['Expires']))
@@ -240,7 +240,7 @@ class TestProfileWithoutCaching(unittest.TestCase):
         # This should use cacheInBrowser
         self.assertEqual('max-age=0, must-revalidate, private',
                          browser.headers['Cache-Control'])
-        tag = '"||{0}|en|{1}|0'.format(catalog.getCounter(), default_skin)
+        tag = '"||{0}|en|{1}'.format(catalog.getCounter(), default_skin)
         self.assertEqual(tag, _normalize_etag(browser.headers['ETag']))
         self.assertGreater(now, dateutil.parser.parse(
             browser.headers['Expires']))
@@ -260,7 +260,7 @@ class TestProfileWithoutCaching(unittest.TestCase):
         self.assertIn(testText, browser.contents)
         self.assertEqual('max-age=0, must-revalidate, private',
                          browser.headers['Cache-Control'])
-        tag = '"||{0}|en|{1}|0'.format(catalog.getCounter(), default_skin)
+        tag = '"||{0}|en|{1}'.format(catalog.getCounter(), default_skin)
         self.assertEqual(tag, _normalize_etag(browser.headers['ETag']))
         self.assertGreater(now, dateutil.parser.parse(
             browser.headers['Expires']))