+++ CHANGES.rst
@@ -10,7 +10,8 @@ Breaking changes:
 
 New features:
 
-- *add item here*
+- Python 3 support
+  [pbauer, MatthewWilkes, ale-rt]
 
 Bug fixes: